Easier Command usage and creation
Added "In Menu" check box (if unchecked, item is not added to context
menu of items, tray menu and "Item" sub-menu).

Copy to tab can be activated from menu too (same as "execute command").

"Ignore Item" renamed to "Remove Item".

Move to tab is combination of "Copy to tab" and "Remove item".