X11: Remove interval while faking Shift+Insert
Automatic paste process is following: For all physically pressed keys a
release event is sent, Shift+Insert is press event is sent and lastly
press event is sent again for the physically pressed keys.

Problem is when user releases some keys during this process. When this
happens, press event is created for keys that are physically not
pressed.

Removing interval, minimizes this issue. This partially reverts commit
b7d5f02 so the automatic paste feature may not work in some apps.