invoke setFocus() and keyEvent()
If the searchbox is active the first item in the list is automatically selected. But if I want to select the second item in the list I have to press Key_Down twice. With this change the first Key_Down will set the focus and also select the second item in the list.