Do not ignore URLs when ignoring copied files
The "Ignore copied files" command ignored all entries that had a
text/uri-list format. Since a URL is a special case of a URI, in
some cases URLs were ignored. For example, when copying a URL from
Chromium's address bar, Chromium (but not Firefox) stores both
text/plain and text/uri-list mime types. These two are the same mime
types that Nautilus stores when copying a file and thus it is
impossible to distinguish between the two based on mime types alone.
Because of this, an extra regex condition is now added to ignore
only items that are not URLs.

It is easier to _not_ match a URL than to match a file in a
cross-platform way, so a negative look ahead regex is used to
exclude URLs.

Note that even if an address is displayed in Chromium's address bar
in simplified form, e.g.:

  hluk.github.io/CopyQ/

if it is copied to the clipboard containing the full form:

  http://hluk.github.io/CopyQ/

and thus the regex correctly matches.