X11: Fix owner when both clipboard and selection
If both clipboard and selection change at the same time, it's better to
postpone fetching data (which can take some time) so as to get correct
clipboard/selection owner (currently active window).

This fixes selection owner for KeePassXC - the app window is hidden
after user name or password are copied.

Signed-off-by: Lukas Holecek <hluk@email.cz>