Windows: Fix blocking key modifiers after paste
In some cases, the state of keyboard modifiers (Ctrl, Shift, Alt, etc.)
was set incorrectly after paste operation. Some of these keys may have
been released by user during the paste, but the application still reset
the key state to pressed programmatically, resulting in incorrect key
state in system.

The fix is not to reset the state afterwards, which can still lead to
incorrect state in system until user releases the keys. User will also
need to release and press the modifiers again (e.g. to be able to use
shortcut to paste) but this is much less annoying.

Fixes #1139