Drop itemweb plugins
The plugin can do some network request to show full web pages in CopyQ.
It may also use a bit outdated libraries. So there is a security
concern.

Apart from that there were some performance issues in past and current
main developer doesn't use the plugin and there are no automated tested.

The plugin was not available in the official Windows builds because it
required some large dependencies to be included.

So, in all, it's better to remove the plugin.

Signed-off-by: Lukas Holecek <hluk@email.cz>