Fix handling of git GPG in Windows (#1689)
* Fix handling of git GPG in Windows

If CopyQ detect's the git bash gpg in path it will use it. However, this version of GPG expects unix paths instead of Windows. Trying to run it with Windows native paths cause it to hang for a lot of time, consuming 100% CPU.

This patch tried to detect whether the given version of GPG in Windows is a unix one, and if such, convert the key paths to unix

* Fix returning a bad value

* Update plugins/itemencrypted/itemencrypted.cpp

Co-authored-by: Lukas Holecek <ruling.worlds@gmail.com>