Simplify window geometry restoring
This changes tries to simplify and improve behavior of window geometry
restore/save functionality.

The main change is that geometry is changed only initially and when
current screen changes. The current screen is indicated by mouse
pointer/cursor position and is applied to the window when the window
"Show" event is triggered. In case the current mouse pointer position
cannot be inspected (for example on some Wayland compositors), it is
left up to the window manager whether it decides to move the window to
another screen (app listens to `QWindow::screenChanged` signals).

This also adds a workaround for resizing window on Sway window
compositor without breaking the window's central position (window
position cannot be set in Qt on most or all Wayland compositors).